change the type of double to float  in the indexedPeak

+++ mzLib/MassSpectrometry/PeakIndexing/IndexedMassSpectralPeak.cs
@@ -6,11 +6,11 @@ namespace MassSpectrometry
     public class IndexedMassSpectralPeak : IIndexedPeak
     {
         public int ZeroBasedScanIndex { get; init; }
-        public double Mz { get; init; }
-        public double M => Mz;
-        public double RetentionTime { get; init; }
-        public double Intensity { get; init; }
-        public IndexedMassSpectralPeak(double mz, double intensity, int zeroBasedMs1ScanIndex, double retentionTime)
+        public float Mz { get; init; }
+        public float M => Mz;
+        public float RetentionTime { get; init; }
+        public float Intensity { get; init; }
+        public IndexedMassSpectralPeak(float mz, float intensity, int zeroBasedMs1ScanIndex, float retentionTime)
         {
             this.Mz = mz;
             this.ZeroBasedScanIndex = zeroBasedMs1ScanIndex;
